fix: move Codex auth merge decision script to static file
